Add a method for adding field after the definition of a serialize
It's useful in case of circular dependency.
Typically we have something like this:

    ASerilizer(Serializer):
        b = BSerializer(many=True)

    BSerializer(Serializer):
        a = ASerializer(many=True)

This isn't possible in python because BSerializer isn't yet define when
we want to use it in ASerilizer. This Function give us the ability to
solve this problem this way:

    ASerilizer(Serializer):
        pass

    BSerializer(Serializer):
        a = ASerializer(many=True)

    ASerilizer.add_field('b', BSerializer(many=True)

Of course if the serialized objects reference themself there will be
ome infinite loop.